# docs/swagger.yaml
basePath: /
definitions:
...
  models.Image:
    properties:
      day_bucket:
        type: string
      image_id:
        type: string
      image_url:
        type: string
      title:
        type: string
      uploaded_at:
        type: string
      user_id:
        type: string
      user_profile_picture_url:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for OSOHUB using Cassandra
  title: OSOHUB API
  version: "1.0"
paths:
...
  /images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image file (JPG, PNG, GIF, WebP, max 10MB)
        in: formData
        name: image
        required: true
        type: file
      - description: Image title (max 100 characters)
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload a new image file to Cloudinary
      tags:
      - Images
  /images/{image_id}:
    delete:
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an image by ID (solo el dueño puede borrar)
      tags:
      - Images
  /images/{image_id}/like:
    delete:
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove like from an image
      tags:
      - Images
    post:
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Like an image (one like per user)
      tags:
      - Images
  /images/{image_id}/likes/count:
    get:
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get like count for an image
      tags:
      - Images
...
  /images/byid/{image_id}:
    get:
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get image by ID (direct, for Swagger compatibility)
      tags:
      - Images
...
  /users/{user_id}/images:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get all images for a user (profile)
      tags:
      - Images
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...